// src/lib/types.ts
export interface JikanCharacter {
    character: {
        mal_id: number;
        url?: string;
        images: {
            jpg: { image_url: string };
            webp: { image_url: string; small_image_url: string };
        };
        name: string;
    };
    role: string;
    favorites?: number;
    voice_actors?: JikanVoiceActor[];
}

export interface JikanVoiceActor {
    person: {
        mal_id: number;
        url?: string;
        images: { jpg: { image_url: string } };
        name: string;
    };
    language: string;
}